// notes.frontend/src/api/api.ts
export interface CreateNoteDto {
  title: string;
  details?: string | undefined;
}